Moved Stats tests to a directory
